taxon matching scripts
- list of taxa extracted from iNat
- scripts to:
  * generate list from full iNat taxonomy (which IS NOT COMPLETE, but a good start)
  * merge together different taxonomic datasets into a single resource
  * revised matching script based off APCalign -- a few fuzzy matches currently "turned off" (for 2-word, 3-word matches)

Notes:
The column "genus" is currently the "1st word" of a string. I'll change that to be more explicit
There are only ~300 taxa that didn't match and relatively few fuzzy matching errors. Of the taxa that don't match, I looked up some in iNat and indeed they don't exist. Some because of name changes, others didn't appear at all.